requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
dbt-duckdb
duckdb
pandas
kagglehub
streamlit
astronomer-cosmos
